requirements.txt
cv2
numpy
soundfile
sounddevice
websockets
wsmprpc
Pillow
